+++ src/app/search-page/page.tsx
@@ -1,30 +1,9 @@
-import { tm } from '@/utils/tw-merge';
-import { useEffect, useRef } from 'react';
-import SearchInput from '@/features/search-page/components/search-page';
+import SearchPageEffect from '@/features/search-page/components/search-effect';
 
 function SearchPage() {
-  const searchFormRef = useRef<{
-    focus: () => void;
-  }>(null);
-
-  useEffect(() => {
-    const clearId = setTimeout(() => {
-      const handles = searchFormRef.current;
-      if (handles) {
-        handles.focus();
-      }
-    }, 900);
-
-    return () => {
-      clearTimeout(clearId);
-    };
-  });
-
   return (
     <>
-      <div className={tm('px-[14px]')}>
-        <SearchInput ref={searchFormRef} />
-      </div>
+      <SearchPageEffect />
     </>
   );
 }
